requirements.txt
# packaging
pip>=9
setuptools>=26
twine
wheel>=0.29

# data science
numpy
pandas
scikit-learn
torch
torchaudio
spectralcluster
python-dotenv

# tests/linter
black
coverage
flake8
pytest
yapf

# API
gcsfs
google-cloud-storage
mlflow
s3fs
uvicorn
fastapi

# utilities
six>=1.14
joblib
memoized-property
termcolor

# pyannote
pyannote.audio
pyannote.database
pyannote.core

#audio
pydub
ffmpeg

#azure
azure.cognitiveservices.speech